Add global UDFs feature
This change consists of the following:

In the API, this adds new endpoints for creating and managing global
UDFs. The global UDFs are stored in a new table, while local UDFs are
still stored in the pipelines table.

In the controller and compiler service, this changes the UDF validation
and compilation logic to include global UDFs. We traverse the job graph
to to determine which UDFs are used, and include only those.

In the console, this adds adds several new components to the create
pipeline page to support browsing local and global UDFs, viewing UDFs in
multiple tabs, and creating and saving new UDFs.